# add Profiler.increment and Profiler.frame