@mzgoddard 2019-07-17
- Update code to fulfill requested changes.