@kchadha 2019-09-16
LGTM but I think I want to get a second pair of eyes on this (either @cwillisf or @rschamp maybe?). I can confirm that I see a large improvement in the benchmark numbers being reported!